# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервера для системы управления научными проектами Scifi.",
        "title": "Scifi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/grants": {
            "get": {
                "description": "Получить список всех грантов с возможностью фильтрации по статусу, типу и поисковому запросу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Список грантов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу (draft, active, completed...)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по типу (state, international...)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск по названию или организации",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объект с массивом 'grants' и числом 'total'",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает запись о гранте или источнике финансирования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Создать новый грант",
                "parameters": [
                    {
                        "description": "Данные гранта",
                        "name": "grant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Grant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Grant"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/grants/{grantId}/projects/{fundingId}": {
            "delete": {
                "tags": [
                    "grants"
                ],
                "summary": "Удалить привязку проекта к гранту",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "grantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID записи финансирования",
                        "name": "fundingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/grants/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Детальная информация о гранте",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Grant"
                        }
                    },
                    "404": {
                        "description": "grant not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "grants"
                ],
                "summary": "Удалить грант",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "description": "Изменить описание, сроки или сумму гранта. Доступно автору или админу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Обновить данные гранта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "grant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Grant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Grant"
                        }
                    },
                    "403": {
                        "description": "insufficient permissions",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/grants/{id}/budget": {
            "get": {
                "description": "Возвращает общую сумму, распределенную сумму и остаток",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Состояние бюджета гранта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объект с полями total, allocated, remaining",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/grants/{id}/projects": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Проекты, финансируемые данным грантом",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProjectGrantFunding"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Выделяет определенную сумму из гранта на конкретный проект",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grants"
                ],
                "summary": "Привязать проект к гранту",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID гранта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные финансирования",
                        "name": "funding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProjectGrantFunding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProjectGrantFunding"
                        }
                    }
                }
            }
        },
...
        "/projects/{id}/grants": {
            "get": {
                "description": "Получить список всех источников финансирования для конкретного проекта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Гранты, привязанные к проекту",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID проекта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ProjectGrantInfo"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Grant": {
            "type": "object",
            "properties": {
                "application_deadline": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "funding_organization": {
                    "type": "string"
                },
                "grant_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "principal_investigator_id": {
                    "type": "integer"
                },
                "scientific_direction": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.ProjectGrantFunding": {
            "type": "object",
            "properties": {
                "allocated_amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "funding_end_date": {
                    "type": "string"
                },
                "funding_purpose": {
                    "type": "string"
                },
                "funding_start_date": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "section_id": {
                    "type": "integer"
                }
            }
        },
        "models.ProjectGrantInfo": {
            "type": "object",
            "properties": {
                "allocated_amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "funding_end_date": {
                    "type": "string"
                },
                "funding_purpose": {
                    "type": "string"
                },
                "funding_start_date": {
                    "type": "string"
                },
                "grant_code": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "integer"
                },
                "grant_title": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "project_id": {
                    "type": "integer"
                },
                "project_key": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "section_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}